CXP-694: Migrated BarChart to typescript
